Rename InteractionDefaults
